Add user data deletion route

+++ domain/src/main/kotlin/com/github/enteraname74/cloudy/domain/model/user/UserStorage.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class UserStorage(
-    val max: Int,
-    val current: Int,
+    val max: Double,
+    val current: Double,
 )
